requirements.txt
asgiref==3.7.2
channels==3.0.4
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3.post1
sqlparse==0.4.4
